+++ src/screens/Dashboard/Profile/index.tsx
@@ -1,7 +1,8 @@
-import { ScreenTitle } from 'components/Common';
+import { ScreenTitle, Title } from 'components/Common';
 import { DashboardScreenContainer } from 'components/Dashboard';
+import { ProfileForm } from 'components/Profile';
 import React, { FC } from 'react';
-import { StyleProps } from 'styles';
+import { colors, mx, StyleProps, Styles } from 'styles';
 
 type Props = StyleProps;
 
@@ -9,9 +10,21 @@ export const DashboardProfileScreen: FC<Props> = () => {
   return (
     <>
       <ScreenTitle title="Profile" />
-      <DashboardScreenContainer>{`Profile`}</DashboardScreenContainer>
+      <DashboardScreenContainer>
+        <Title style={styles.title} type="h1">
+          {'Profile'}
+        </Title>
+        <ProfileForm />
+      </DashboardScreenContainer>
     </>
   );
+};
+
+const styles: Styles = {
+  title: {
+    ...mx.font(50, colors.marineBlue),
+    marginBottom: 38,
+  },
 };
 
 export type DashboardProfileScreenProps = Props;